Initial work on getting values from a log scale axix.  currently hardcoded so that Y axis is always log. also, axis start and end is hard coded atm due to a bug with floating point input values from the GUI

+++ GraphValues/DataPoint.cs
@@ -9,10 +9,10 @@ namespace GraphValues
 {
     class DataPoint
     {
-        public float X { private set; get; }
-        public float Y { private set; get; }
+        public double X { private set; get; }
+        public double Y { private set; get; }
 
-        public DataPoint(float x, float y)
+        public DataPoint(double x, double y)
         {
             X = x;
             Y = y;
